requirements.txt
requests>=2.28
PyPDF2>=3.0
python-docx>=0.8
openpyxl>=3.0
python-pptx>=0.6
blessed>=1.20
pyyaml>=6.0
